+++ cloudauth-20190307/alibabacloud_cloudauth20190307/models/__init__.py
@@ -240,6 +240,9 @@ from ._mobile_online_status_response import MobileOnlineStatusResponse
 from ._mobile_online_time_request import MobileOnlineTimeRequest
 from ._mobile_online_time_response_body import MobileOnlineTimeResponseBody
 from ._mobile_online_time_response import MobileOnlineTimeResponse
+from ._mobile_recycled_meta_verify_request import MobileRecycledMetaVerifyRequest
+from ._mobile_recycled_meta_verify_response_body import MobileRecycledMetaVerifyResponseBody
+from ._mobile_recycled_meta_verify_response import MobileRecycledMetaVerifyResponse
 from ._modify_black_list_strategy_request import ModifyBlackListStrategyRequest
 from ._modify_black_list_strategy_shrink_request import ModifyBlackListStrategyShrinkRequest
 from ._modify_black_list_strategy_response_body import ModifyBlackListStrategyResponseBody
@@ -394,6 +397,7 @@ from ._mobile_detect_response_body import MobileDetectResponseBodyResultObjectIt
 from ._mobile_detect_response_body import MobileDetectResponseBodyResultObject
 from ._mobile_online_status_response_body import MobileOnlineStatusResponseBodyResultObject
 from ._mobile_online_time_response_body import MobileOnlineTimeResponseBodyResultObject
+from ._mobile_recycled_meta_verify_response_body import MobileRecycledMetaVerifyResponseBodyResultObject
 from ._modify_black_list_strategy_request import ModifyBlackListStrategyRequestBlackListStrategy
 from ._modify_control_strategy_request import ModifyControlStrategyRequestControlStrategyList
 from ._modify_customize_flow_strategy_list_request import ModifyCustomizeFlowStrategyListRequestStrategyObject
@@ -653,6 +657,9 @@ __all__ = [
     MobileOnlineTimeRequest,
     MobileOnlineTimeResponseBody,
     MobileOnlineTimeResponse,
+    MobileRecycledMetaVerifyRequest,
+    MobileRecycledMetaVerifyResponseBody,
+    MobileRecycledMetaVerifyResponse,
     ModifyBlackListStrategyRequest,
     ModifyBlackListStrategyShrinkRequest,
     ModifyBlackListStrategyResponseBody,
@@ -807,6 +814,7 @@ __all__ = [
     MobileDetectResponseBodyResultObject,
     MobileOnlineStatusResponseBodyResultObject,
     MobileOnlineTimeResponseBodyResultObject,
+    MobileRecycledMetaVerifyResponseBodyResultObject,
     ModifyBlackListStrategyRequestBlackListStrategy,
     ModifyControlStrategyRequestControlStrategyList,
     ModifyCustomizeFlowStrategyListRequestStrategyObject,